+++ src/icons/transportation/index.ts
@@ -1,0 +1,8 @@
+/**
+ * This file exports all icons in the "Transportation" category.
+ */
+import Airplane from '@/src/icons/transportation/Airplane';
+import Bus from '@/src/icons/transportation/Bus';
+import Cab from '@/src/icons/transportation/Cab';
+
+export { Airplane, Bus, Cab };